[config] Add network related configs. JB#60548
This adds the relevant network configs to be checked. Set to require
most of the basic needs for networking. Some "good to have" are marked
as optional.

This utilizes the new approach for allowing to check if the kernel
version is in range. The config NF_NAT_IPV{4,6} exists only in kernels
between 3.7 and 5.0 and works as an example.